requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.0
ffmpeg==1.4
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.100.0
openai-whisper==20250625
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
